Use rowids instead of composite keys
The composite keys worked, but then queries to retrieve specific
words, synsets, or senses would need the id, lexicon id, and lexicon
version. Using rowids means that the user-facing objects will store
DB-internal data (the rowids), although it is hidden from the user.

This commit also changes the get_sense() function to find_senses(),
parallel to the find_entries() and find_synsets() functions.

The maps in wn._db are gone, too. There seemed to be no performance
benefit, although now the queries are slightly more complicated.

Finally, this commit takes some data out of the database (for now):
- lexicographer file on synsets
- sense_id on definitions
- sense_key on senses
- adjposition_id on senses

These were taken out because they are not applicable to all wordnets
or even all elements, so they may find a new home in another
table. The 'lexicalized' attribute on synsets and senses might also be
moved, too, as it is almost never used.