Keep wordnet instance reference on other entities
When searching wn.synsets(), wn.words(), or wn.senses(), the returned
objects now have a reference to the WordNet object used when
searching. As part of this change, the module functions now create a
WordNet object with the `lgcode` and `lexicon` arguments. This is to
help with calls such as Synset.hypernyms() to look within the same
lexicons.