Allow Synset.get_related() to return empty synsets
Also much of the complexity regarding retrieving synset relations is
now moved to the Python code where it is more natural, and the SQL
code is greatly simplified. This change also needed a new _db function
to retrieve synsets from a collection of ILIs (actually it's not
necessary, but this does in one query what otherwise takes many
separate, similar queries).