Streamline _db code a bit and return more info
This commit replaces similar DB queries with dynamic queries that
select the table from a parameter. To avoid SQL injection, the table
parameter is sanitized through dictionary lookup (so the user-supplied
value is never used in the formatted queries, only predefined strings
are). This commit also changes the query functions for examples and
definitions to return their other columns (including rowid), which
will help with retrieving things like metadata later. Unlike the
previous changes to word form, these other columns are not available
to the user through the public API.