Fix #106: add wn.DatabaseError for schema changes
The error is named DatabaseError instead of SchemaError or even
IncompatibleSchemaError so that it might be used for other database
problems, if they are found and caught. Coincidentally, the exception
hierarchy is now mirroring Python's sqlite3 module's hierarchy, but
thi is not intentional and there is no plan to continue that trend.