Back off failed lemmatization to original form
This is to account for the (currently hypothetical) case where a user
searches for a word with a part of speech not covered by the
lemmatizer, but where the word exists in the database. For example,
the Morphy lemmatizer only covers nouns, verbs, adjectives, and
adverbs, so if an English wordnet ever includes other parts of
speech (prepositions, conjunctions, phrases, etc.) it would yield
nothing and queries would therefore return nothing. Now, if the
lemmatizer yields nothing, the original form and part of speech are
used.