Simplify wn.morphy and lemmatization
I'm revising the plan for lemmatization such that the Wordnet class
takes only a callable for the `lemmatizer` parameter and doesn't
perform any instantiation of the lemmatizer. The `Lemmatizer` class is
therefore unnecessary and is removed. Morphy now has two modes:
uninitialized and initialized. Initialized is as before, but
uninitialized only performs the de-suffixing and does not validate the
lemmas. As such, it is more suitable for applications requiring fast
startup, such as web servers. Also, lemmatizer functions now return a
mapping of parts of speech to sets of lemmas instead of yielding
lemmas only. This allows it to take `None` for a part of speech and do
something reasonable.