Fix #113: reduce scope of lmf.scan_lexicons()
Now it only looks for lexicons and gets their IDs and versions in
order to verify if they are already added or have dependencies
satisfied. Counting for progress-reporting happens on the lexicon
object.

Before, scan_lexicons() took about 2.5s and the element counting took
a negligible amount of time. Now, scan_lexicons takes about .4s and
counting takes about .5s. This means it's about a second faster to add
a lexicon, but about 2x as fast to abort the adding of an
already-added lexicon.