Model relations in _Relatable.incoming_relation()
Senses and Synsets now have an `incoming_relation()` method. The value
of this method returns a SenseRelation, SenseSynsetRelation, or
SynsetRelation object if the Sense/Synset is the result of a relation
traversal. Otherwise the method returns `None`.

The new relation objects specify the relation name, the source and
target IDs of the relation, and the lexicon where the relation
originated.

Fixes #216
Fixes #167