Use simpler .relation_map() method
The .relation_map() method is like .relations(), but the keys are
relation objects and map 1-to-1 to targets.

This commit also removes the .source() and .target() methods as they
didn't work in common situations, as well as the .incoming_relation()
method as it was not favored.

This commit also adds a new utility function and associated unit tests.