Add docs and dependencies related to benchmarking
With this the benchmarking support should be ready to go. The
benchmarks themselves are not exhaustive; rather I just tried to
create the benchmarking framework and solve some of the problems that
came up with a few tests. When making changes to some part of Wn, the
author may wish to create a relevant benchmark to test before/after
speeds.

Resolves #98